udated the teams module

+++ bizlaunch/users/views.py
@@ -5,14 +5,23 @@ from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework import status, viewsets
 from rest_framework.decorators import action
+from rest_framework.parsers import FormParser, MultiPartParser
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework_simplejwt.tokens import RefreshToken
 
-from bizlaunch.users.models import Profile, Team, TeamInvite, TeamMember
-from bizlaunch.users.permissions import IsProfileOwner, IsTeamOwner
+from bizlaunch.users.models import (
+    InviteStatus,
+    Profile,
+    Team,
+    TeamInvite,
+    TeamMember,
+    UserRole,
+)
+from bizlaunch.users.permissions import IsProfileOwner
 from bizlaunch.users.serializers import (
+    ChangePasswordSerializer,
     MemberRegisterSerializer,
     ProfileSerializer,
     TeamInviteSerializer,
@@ -33,6 +42,7 @@ def password_reset_confirm_redirect(request, uidb64, token):
 
 class ProfileView(APIView):
     permission_classes = [IsAuthenticated, IsProfileOwner]
+    parser_classes = [MultiPartParser, FormParser]
 
     @swagger_auto_schema(
         operation_description="Retrieve the authenticated user's profile",
@@ -48,7 +58,21 @@ class ProfileView(APIView):
 
     @swagger_auto_schema(
         operation_description="Partially update the authenticated user's profile",
-        request_body=ProfileSerializer,
+        manual_parameters=[
+            openapi.Parameter(
+                "bio",
+                openapi.IN_FORM,
+                description="User bio",
+                type=openapi.TYPE_STRING,
+            ),
+            openapi.Parameter(
+                "profile_picture",
+                openapi.IN_FORM,
+                description="Profile picture file upload",
+                type=openapi.TYPE_FILE,
+            ),
+        ],
+        consumes=["multipart/form-data"],  # Specify the content type for file uploads
         responses={
             200: ProfileSerializer,
             400: openapi.Response(description="Invalid input"),
@@ -68,26 +92,6 @@ class ProfileView(APIView):
         )
 
 
-# class TeamInviteCreateView(APIView):
-#     permission_classes = [IsAuthenticated]
-#     serializer_class = TeamInviteSerializer
-
-#     @swagger_auto_schema(
-#         operation_description="Create a team invite for a user",
-#         request_body=TeamInviteSerializer,
-#         responses={
-#             201: TeamInviteSerializer,
-#             400: openapi.Response(description="Invalid input"),
-#             403: openapi.Response(description="Permission denied"),
-#         },
-#     )
-#     def post(self, request, *args, **kwargs):
-#         serializer = self.serializer_class(data=request.data, context={"request": request})
-#         serializer.is_valid(raise_exception=True)
-#         serializer.save(inviter=request.user)
-#         return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-
 class TeamViewSet(viewsets.ViewSet):
     permission_classes = [IsAuthenticated]
     lookup_field = "uuid"
@@ -97,52 +101,27 @@ class TeamViewSet(viewsets.ViewSet):
         return Team.objects.filter(owner=self.request.user)
 
     @swagger_auto_schema(
-        operation_description="Retrieve the authenticated user's team",
+        operation_description="Retrieve the authenticated user's team. If no team exists, one will be created automatically. The response includes team details, pending invites, and accepted members.",
         responses={status.HTTP_200_OK: TeamSerializer},
     )
-    def list(self, request):
+    def get(self, request):
+        # Check if the user owns a team
         team = self.get_queryset().first()
         if not team:
-            return Response(
-                {"detail": "You do not own a team."}, status=status.HTTP_404_NOT_FOUND
-            )
+            # Automatically create a team for the user if they are an admin
+            if request.user.role == UserRole.ADMIN:
+                team = Team.objects.create(
+                    owner=request.user, name=f"{request.user.name}'s Team"
+                )
+            else:
+                return Response(
+                    {"detail": "You are not authorized to own a team."},
+                    status=status.HTTP_403_FORBIDDEN,
+                )
+
+        # Serialize the team, including members and invites
         serializer = TeamSerializer(team)
         return Response(serializer.data)
-
-    @swagger_auto_schema(
-        operation_description="Create a new team for the authenticated user",
-        request_body=openapi.Schema(
-            type=openapi.TYPE_OBJECT,
-            properties={
-                "name": openapi.Schema(
-                    type=openapi.TYPE_STRING, description="Name of the team"
-                ),
-            },
-            required=["name"],
-        ),
-        responses={
-            status.HTTP_201_CREATED: TeamSerializer,
-            status.HTTP_400_BAD_REQUEST: openapi.Response(
-                description="Team already exists"
-            ),
-        },
-    )
-    def create(self, request):
-        if self.get_queryset().exists():
-            return Response(
-                {"detail": "You already own a team."},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        team_name = request.data.get("name")
-        if not team_name:
-            return Response(
-                {"detail": "Team name is required."}, status=status.HTTP_400_BAD_REQUEST
-            )
-
-        team = Team.objects.create(owner=request.user, name=team_name)
-        serializer = TeamSerializer(team)
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     @swagger_auto_schema(
         operation_description="Update the authenticated user's team",
@@ -160,27 +139,6 @@ class TeamViewSet(viewsets.ViewSet):
         serializer.save()
         return Response(serializer.data)
 
-    @swagger_auto_schema(
-        operation_description="Delete the authenticated user's team",
-        responses={status.HTTP_204_NO_CONTENT: "No Content"},
-    )
-    def destroy(self, request, pk=None):
-        team = self.get_queryset().first()
-        if not team:
-            return Response(
-                {"detail": "You do not own a team."}, status=status.HTTP_404_NOT_FOUND
-            )
-        team.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-    @swagger_auto_schema(
-        operation_description="List team members segregated by invite status",
-        responses={
-            status.HTTP_200_OK: openapi.Response(
-                "Team members", TeamMemberSerializer(many=True)
-            )
-        },
-    )
     @action(detail=False, methods=["get"], url_path="members")
     def list_members(self, request):
         team = self.get_queryset().first()
@@ -206,6 +164,11 @@ class TeamViewSet(viewsets.ViewSet):
     )
     @action(detail=False, methods=["post"], url_path="invite-member")
     def invite_member(self, request):
+        if request.user.role != UserRole.ADMIN:
+            return Response(
+                {"detail": "Only team admins can invite members."},
+                status=status.HTTP_403_FORBIDDEN,
+            )
         serializer = TeamInviteSerializer(
             data=request.data, context={"request": request}
         )
@@ -216,24 +179,46 @@ class TeamViewSet(viewsets.ViewSet):
         )
 
     @swagger_auto_schema(
-        operation_description="Delete a team member by UUID",
+        operation_description="Delete a team member or team invite by UUID",
         responses={status.HTTP_204_NO_CONTENT: "No Content"},
     )
-    @action(detail=True, methods=["delete"], url_path="members/(?P<member_uuid>[^/.]+)")
-    def delete_member(self, request, member_uuid=None, uuid=None):
+    @action(detail=False, methods=["delete"], url_path="member/(?P<uuid>[^/.]+)")
+    def delete_member_or_invite(self, request, uuid=None):
+        if request.user.role != UserRole.ADMIN:
+            return Response(
+                {"detail": "Only team admins can remove members or invites."},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
         team = self.get_queryset().first()
         if not team:
             return Response(
                 {"detail": "You do not own a team."}, status=status.HTTP_404_NOT_FOUND
             )
 
+        # Attempt to delete a team member
         try:
-            member = TeamMember.objects.get(uuid=member_uuid, team=team)
+            member = TeamMember.objects.get(uuid=uuid, team=team)
             member.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         except TeamMember.DoesNotExist:
+            pass  # If not found, proceed to check for invites
+
+        # Attempt to delete a pending or expired team invite
+        try:
+            invite = TeamInvite.objects.get(uuid=uuid, team=team)
+            if invite.status in [InviteStatus.PENDING, InviteStatus.EXPIRED]:
+                invite.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            else:
+                return Response(
+                    {"detail": "Only pending or expired invites can be deleted."},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
+        except TeamInvite.DoesNotExist:
             return Response(
-                {"detail": "Team member not found."}, status=status.HTTP_404_NOT_FOUND
+                {"detail": "Team member or invite not found."},
+                status=status.HTTP_404_NOT_FOUND,
             )
 
 
@@ -278,4 +263,34 @@ class MemberRegisterView(APIView):
                 "refresh_token": refresh_token,
             },
             status=status.HTTP_201_CREATED,
+        )
+
+
+class ChangePasswordView(APIView):
+    permission_classes = [IsAuthenticated]
+
+    @swagger_auto_schema(
+        operation_description="Change the password of the authenticated user",
+        request_body=ChangePasswordSerializer,
+        responses={
+            200: openapi.Response(
+                description="Password updated successfully",
+                examples={
+                    "application/json": {"detail": "Password updated successfully."}
+                },
+            ),
+            400: openapi.Response(description="Invalid input"),
+            401: openapi.Response(
+                description="Authentication credentials were not provided."
+            ),
+        },
+    )
+    def post(self, request, *args, **kwargs):
+        serializer = ChangePasswordSerializer(
+            data=request.data, context={"request": request}
+        )
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+        return Response(
+            {"detail": "Password updated successfully."}, status=status.HTTP_200_OK
         )
